v7/morando/requirements.txt
cached-property==1.3.0
coreapi==2.0.8
Django==1.10.2
django-auto-rest-project==0.4.1
django-braces==1.9.0
django-cors-headers==1.2.2
django-drf-file-generator==0.1.13
django-filter==0.15.3
django-jet==1.0.2
django-oauth-toolkit==0.10.0
django-rest-auth==0.8.1
django-rest-framework-social-oauth2==1.0.4
django-rest-swagger==2.0.7
django-social-auth==0.7.28
django-url-filter==0.2.0
djangorestframework==3.5.1
enum34==1.1.6
feedparser==5.2.1
google-api-python-client==1.5.4
httplib2==0.9.2
itypes==1.1.0
Markdown==2.6.7
MySQL-python==1.2.5
oauth2==1.9.0.post1
oauth2client==4.0.0
oauthlib==1.0.3
openapi-codec==1.1.7
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.11.1
requests-oauthlib==0.7.0
rsa==3.4.2
simplejson==3.9.0
six==1.10.0
uritemplate==3.0.0
